+++ src/lib.rs
@@ -34,23 +34,21 @@ pub fn parse_arguments(arguments: Vec<String>, env: Env) {
         // is always the first element of env::args
         0 => panic!("Something went horribly wrong."),
         1 => print_standup_data(&file),
-        2 => {
-            let default_editor =
-                env::var("EDITOR").expect("No $EDITOR environment variable found.");
-            match arguments[1].as_str() {
-                CLEAR => data_file::clear_data_from_file(&file),
-                EDIT => data_file::manually_edit_file(&file, default_editor),
-                HELP | DASH_HELP => print_help_information(),
-                _ => print_invalid_command(),
+        2 => match arguments[1].as_str() {
+            CLEAR => data_file::clear_data_from_file(&file),
+            EDIT => {
+                let default_editor = env::var("EDITOR").unwrap_or_else(|_| "vi".to_string());
+                data_file::manually_edit_file(&file, default_editor)
             }
-        }
+            HELP | DASH_HELP => print_help_information(),
+            _ => print_invalid_command(),
+        },
         3 => {
             let command = arguments[1].as_str();
             let user_input = arguments[2].as_str();
             let standup = data_file::read_from_file(&file);
             match command {
                 DID | DI | DOING | DO | BLOCKER | BL | SIDEBAR | SB => {
-                    // TODO! Fix Using Moved Value error below
                     standup.add_item(&file, command, user_input)
                 }
                 EDIT => data_file::manually_edit_file(&file, user_input.to_string()),