+++ Changelog.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-**Empty**
+### Fixed
+- App would crash if `EDITOR` environment variable is not set
 
 ## [1.1.0] - 2022-04-26
 ### Added